pyproject.toml
[project]
name = "ai-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.13.2",
    "aiosqlite>=0.21.0",
    "boto3>=1.41.2",
    "gradio>=5.50.0",
    "openai>=2.8.1",
    "peewee>=3.18.3",
    "python-dotenv>=1.2.1",
    "termcolor>=3.2.0",
    "tqdm>=4.67.1",
]
